requirements.txt
fastapi
uvicorn
pymongo
motor
python-dotenv
pydantic
pydantic-settings
openai
faiss-cpu
numpy
tiktoken